+++ EditorDialogUtil.c
@@ -449,6 +449,8 @@ BOOL DoDlgInitPars(HWND hDlg, DIALPARS *dp, int nParams)
 				SetDlgItemText(hDlg,item,(LPSTR)ip);
 				break;
 			case IDD_PATH1:
+				hist_fillComboBox(hDlg, item, PATHES);
+				// drop through
 			case IDD_STRING1: case IDD_STRING2: case IDD_STRING3:
 			case IDD_STRING4: case IDD_STRING5: case IDD_STRING6:
 			case IDD_STRING7: